pyproject.toml
[tool.poetry]
name = "workshop_checker"
version = "0.1.0"
description = "Arma Workshop mod update checker for Zeusops"
authors = ["Sami Laine <sami.v.laine@gmail.com>"]
readme = "README.md"
license = "GPL-3.0-or-later"

[tool.poetry.scripts]
workshop-checker = "workshop_checker.cli:cli"
update-db = "workshop_checker.update_db:cli"
check-update = "workshop_checker.check_update:cli"

[tool.poetry.dependencies]
python = "^3.10"
steamfiles = { git = "https://github.com/WinterPhoenix/steamfiles", branch = "patch-1" }
discord-webhook = "*"
# JSON-serializable content
pydantic = "^1.10.2"


# Note: Linters not defined in this file but .pre-commit-config.yaml, which
# installs/manages each tool in its own isolated virtualenv

[tool.poetry.group.test.dependencies]
pytest = "7.*"
# Test coverage
pytest-cov = "3.*"
# Clearer assertion failures (colorful diff)
pytest-clarity = "1.*"
# Expose unittest.Mocker in pytest
pytest-mock = "*"
# Yaml parser
pyyaml = "*"
# Shared test data directory
pytest-datadir = "*"

[tool.poetry.group.docs.dependencies]
# Main documentation-as-code (HTML/PDF generator)
Sphinx = "5.*"
# Read the docs (pretty) theme for sphinx
sphinx-rtd-theme = "1.*"
# Markdown parser for sphinx: Sphinx uses "RST" files, this adds markdown
myst-parser = "*"
# Automatic Python module docs (javadoc-style)
sphinx-autoapi = "*"

[tool.poetry.group.dev.dependencies]
pre-commit = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Make isort work with Black
# Avoids conflicting imports
# As per https://pycqa.github.io/isort/docs/configuration/black_compatibility/#using-a-config-file-such-as-isortcfg
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 79

[tool.black]
line-length = 79

# Avoid pointless warning about performance hit of f-string in loggers
[tool.pylint.message_control]
disable = ["logging-fstring-interpolation"]

[tool.pytest.ini_options]
addopts = """-vv \
      --doctest-modules \
      --cov=workshop_checker \
      --cov-report=xml:test_results/coverage.xml \
      --cov-report=html:test_results/coverage.html \
      --cov-report=term \
      --junit-xml=test_results/results.xml"""

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
